removing 64 bits warnings

+++ iOS/G3MiOSSDK/Commons/Basic/GenericQuadTree.cpp
@@ -16,8 +16,8 @@
 #pragma mark NODE
 
 GenericQuadTree_Node::~GenericQuadTree_Node() {
-  const int elementsSize = _elements.size();
-  for (int i = 0; i < elementsSize; i++) {
+  const size_t elementsSize = _elements.size();
+  for (size_t i = 0; i < elementsSize; i++) {
     delete _elements[i];
   }
 
@@ -62,8 +62,8 @@ void GenericQuadTree_Node::splitNode(int maxElementsPerNode,
 
   //Split elements by sector
   std::vector<GenericQuadTree_Element*> elementsToBeInserted = _elements;
-  const int size = elementsToBeInserted.size();
-  for (int i = 0; i < size; i++) {
+  const size_t size = elementsToBeInserted.size();
+  for (size_t i = 0; i < size; i++) {
     this->add(elementsToBeInserted[i], maxElementsPerNode, maxDepth, childAreaProportion);
   }
   _elements.clear();
@@ -221,7 +221,7 @@ bool GenericQuadTree_Node::acceptVisitor(const Sector& sector,
     return false;
   }
 
-  const int elementsSize = _elements.size();
+  const size_t elementsSize = _elements.size();
   visitor.addComparisonsDoneWhileVisiting(elementsSize);
 
   for (int i = 0; i < elementsSize; i++) {
